Refactoring a testhelper, to avoid running too many tests.
