+++ tests/services/BggGameService.test.ts
@@ -1,7 +1,7 @@
 import BggGameService from "../../src/services/BggGameService";
 import * as fetchMock from "fetch-mock";
 import { readFileSync } from "fs";
-import { GameInfo } from "../../src/models/GameInfo";
+import { getLargeCollection } from "./TestHelpers";
 
 const proxyUrl = "https://cors-anywhere.herokuapp.com/";
 const expectedUrl = `${proxyUrl}https://api.geekdo.com/xmlapi2/collection?username=Warium&own=1&stats=1&excludesubtype=boardgameexpansion`;
@@ -255,37 +255,3 @@ describe("BggGameService", () => {
     });
 
 });
-
-let largeCollectionCache: GameInfo[] = undefined;
-export async function getLargeCollection(fetch: fetchMock.FetchMockSandbox = fetchMock.sandbox()) {
-    if (largeCollectionCache === undefined) {
-        const service = new BggGameService(fetch);
-        const largeCollection = readFileSync("tests/services/testxml/TheJadeKnightCollection.xml", "utf8");
-        fetch.mock(expectedUrl, 200, {
-            response: {
-                body: largeCollection
-            },
-            overwriteRoutes: true
-
-        });
-        largeCollectionCache = await service.getUserCollection("Warium") as GameInfo[];
-    }
-    return largeCollectionCache as GameInfo[];
-}
-
-let hugeCollection: GameInfo[] = undefined;
-export async function getHugeCollection(fetch: fetchMock.FetchMockSandbox = fetchMock.sandbox()) {
-    if (hugeCollection === undefined) {
-        const service = new BggGameService(fetch);
-        const largeCollection = readFileSync("tests/services/testxml/TomVasel.xml", "utf8");
-        fetch.mock(expectedUrl, 200, {
-            response: {
-                body: largeCollection
-            },
-            overwriteRoutes: true
-        });
-        hugeCollection = await service.getUserCollection("Warium") as GameInfo[];
-    }
-    return hugeCollection as GameInfo[];
-}
-
